Refactor from implict suggestions to explict ones
The extended autocomplete component of users had to implement a computed property `suggestions`. This computed property had two dependencies on the internals of the component which introduces unnecessary coupling and makes future development of the component harder.

Now the extended autocomplete component just has  to implement a method with a specific signature which is called whenever the suggestions need to be determined.