Add test case to expose clearing quirk
Whenever the input is blurred the component tries to determine if the input should be cleared. This happens unless the input text matches one option exactly. While the options get lowercased before comparing the input doesn't. This results in the weird behaviour that `ABC` and `abc` will keep the input for example while `Abc` doesn't.